refactor: create component for user address input
